refactored id form Integer to Long for method findById

+++ src/main/java/com/songifyDatabase/song/domain/repository/SongRepository.java
@@ -11,5 +11,5 @@ public interface SongRepository extends Repository<Song, Long> {
 
     List<Song> findAll();
 
-    Optional<Song> findById(Integer id);
+    Optional<Song> findById(Long id);
 }
